backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai==0.3.0

# Auth
PyJWT>=2.8
cryptography>=42
httpx>=0.27

# Persistence
psycopg[binary]>=3.2
psycopg-pool>=3.2

# Cloud storage
boto3>=1.34

# Cache / rate limiting
redis[hiredis]>=5
# faster-whisper is installed from the cloned repo at ../vendor/faster-whisper
# (see backend/README.md). Its own runtime deps:
ctranslate2>=4.0,<5
huggingface-hub>=0.13
tokenizers>=0.13,<1
onnxruntime>=1.14,<2
av>=11
tqdm
